renamed _b2v4 to b2v4, b2 points to b2v4
