+++ b2/_internal/b2v3/registry.py
@@ -9,7 +9,7 @@
 ######################################################################
 
 # ruff: noqa: F405
-from b2._internal._b2v4.registry import *  # noqa
+from b2._internal.b2v4.registry import *  # noqa
 from b2._internal._cli.b2api import _get_b2api_for_profile
 from b2._internal.arg_parser import enable_camel_case_arguments
 from .rm import Rm
