+++ b2/_internal/b2v4/__init__.py
@@ -1,6 +1,6 @@
 ######################################################################
 #
-# File: b2/_internal/_b2v4/__init__.py
+# File: b2/_internal/b2v4/__init__.py
 #
 # Copyright 2023 Backblaze Inc. All Rights Reserved.
 #
