+++ changelog.d/+apiver-v4.changed.md
@@ -1,0 +1,1 @@
+CLI `b2` executable now points to the latest stable ApiVer version, `b2v4`, previously it pointed to `b2v3`.
