+++ pyproject.toml
@@ -65,9 +65,9 @@ full = [
 Homepage = "https://github.com/Backblaze/B2_Command_Line_Tool"
 
 [project.scripts]
-b2 = "b2._internal.b2v3.__main__:main"
+b2 = "b2._internal.b2v4.__main__:main"
 b2v3 = "b2._internal.b2v3.__main__:main"
-_b2v4 = "b2._internal._b2v4.__main__:main"
+b2v4 = "b2._internal.b2v4.__main__:main"
 
 [build-system]
 requires = ["pdm-backend"]
